Added matchbox (MFJ16010) "personality" file
Added the MFJ 16010 "personality" file.    This we obtained once by collecting S11, S12, S21 and S22 for all combinations of settings.   Note S21 and S12 should be the same (this is true for any "reciprocal network" which passive tuners are) so you can save yourself some work!    Also remember that when you collect S11 or S22 you must terminate the other port with 50 ohms resistance.